adding a working review component with react
