+++ client/components/index.js
@@ -8,3 +8,4 @@ export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
 export {default as AllProducts} from './all-products'
 export {default as SingleProduct} from './SingleProduct'
+export {default as Review} from './Review'
