+++ client/routes.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, Route, Switch, Link} from 'react-router-dom'
 import PropTypes from 'prop-types'
-import {Login, Signup, UserHome, AllProducts, SingleProduct} from './components'
+import {Login, Signup, UserHome, AllProducts, SingleProduct, Review} from './components'
 import {me} from './store'
 
 /**
@@ -22,7 +22,7 @@ class Routes extends Component {
         <Route path="/login" component={Login} />
         <Route path="/signup" component={Signup} />
         <Route exact path="/product" component={AllProducts} />
-        <Route exact path="/product/:productId" component={SingleProduct} />
+        <Route exact path="/product/:productId" component={Review} />
         {isLoggedIn && (
           <Switch>
             {/* Routes placed here are only available after logging in */}
